fix(accounts): probe direct-API keys before keeping them in rotation (#11033 regression)
PR #11033's verify-before-evict short-circuits on any truthy token, but for
DIRECT-API providers (anthropic-api/openai-api/cerebras-api) getAccessToken
resolves the stored key offline — its `expires` is the never-expires sentinel —
so isSubscriptionProvider is false, no server probe runs, and the code logs
"credential verifies — keeping it in rotation" WITHOUT verifying anything. A
revoked direct-API key that 401'd a sub-agent stays health=ok forever, driving
doomed failover respawns onto it. On develop the mark was unconditional, so this
was a strict regression for direct-API pools.

Fix: for direct-API providers, run a real server-side probe (GET /models) before
keeping the key. Only a 2xx keeps it; a 401/403 falls through to
markNeedsReauth; an inconclusive result (network blip, status 0) is left to the
keep-alive sweep — matching the subscription path's transient-blip handling.

- Extract probeDirectApiKey + directProviderBaseUrl from accounts-routes.ts into
  a shared @elizaos/agent/auth/direct-api-probe module (accounts-routes re-imports
  them; behavior unchanged there).
- Add isDirectAccountProvider type guard next to isSubscriptionProvider.
- Wire the probe into coding-account-bridge's markNeedsReauth.

Tests (extend multi-account-refresh-failover.test.ts, real pool + stubbed fetch,
no mock of the bridge): a revoked direct-API key (probe 401) → needs-reauth; a
valid one (probe 200) → kept; an inconclusive probe (status 0) → left alone.
Fail-without-fix proven: reverting the bridge change reds the revoked-key test.
Both packages typecheck exit 0.
